Push to prod: relocate `/docs/topics` to `/topics`

+++ build/init-topics.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 # See README for usage
-# This file will create stub files for all the docs/topics.
+# This file will create stub files for all the topics.
 
 # first check to make sure there are arguments
 if [ -z "$1" ]; then
@@ -14,28 +14,37 @@ if [ ! -d "$1" ]; then
     exit 1
 fi
 
+# check for old style /docs/topics
+if [ -e "content/docs/topics" ]; then
+    echo "Documentation topic files have moved. Delete content/docs/topics before proceeding."
+    exit 1
+fi
 
-ln -s $1 ./build-topics 
+# Create symlink to topics, except if it already exists and points to the same directory.
+if [ ! -L build-topics -o "$(readlink build-topics)" != "$1" ]; then
+    ln -s $1 ./build-topics
+fi
 
 for fname in $(find $1 -maxdepth 1  -iname "*.md")
 do
     base=${fname##*/}
     topic=${base%.*}
-    if [[ "$topic" != "index" ]]; then 
-        cat << EOF > "./content/docs/topics/$topic.md"
+    if [[ "$topic" != "index" ]]; then
+        cat << EOF > "./content/topics/$topic.md"
 +++
 # This is a generated stub file.
 # To edit the content see /topic/$topic.md in the 'valkey-doc' repo
+aliases = ["/docs/topics/$topic/"]
 +++
 EOF
 fi
 done
 
-echo "Topic stub files created."
+echo "Topic stub files created at content/topics."
 
 for fname in $(find $1 -maxdepth 1  -iname "*.png" -o -iname "*.jpg" -o -iname "*.jpeg" -o -iname "*.gif")
 do
     base=${fname##*/}
-    cp ${fname} ./content/docs/topics/${base}
+    cp ${fname} ./content/topics/${base}
 done
 echo "Copied images to topics directory."
